feat(playground): migrate E2E suite to packages/untp-playground/e2e/
