+++ e2e/cypress/support/e2e.ts
@@ -1,3 +1,2 @@
 import './commands/api-auth';
-import './commands/untp-playground';
 import './commands/common';
