+++ packages/untp-playground/e2e/cypress/e2e/report-generation.cy.ts
@@ -1,4 +1,4 @@
-import { TestCaseStatus } from '../../../../packages/untp-playground/constants';
+import { TestCaseStatus } from '../../../constants';
 import { config } from '../../support/config';
 
 describe('Report Generation', () => {
