+++ packages/untp-playground/e2e/cypress/support/e2e.ts
@@ -1,0 +1,1 @@
+import './commands/untp-playground';
