+++ packages/untp-playground/package.json
@@ -9,7 +9,9 @@
     "lint": "next lint",
     "test": "jest",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage"
+    "test:coverage": "jest --coverage",
+    "test:e2e": "cypress run --project ./e2e",
+    "test:e2e:open": "cypress open --project ./e2e"
   },
   "dependencies": {
     "@radix-ui/react-dialog": "^1.1.6",
@@ -43,6 +45,7 @@
     "@types/node": "^20",
     "@types/react": "19.0.4",
     "@types/react-dom": "19.0.4",
+    "cypress": "^14.0.2",
     "eslint": "^8",
     "eslint-config-next": "15.0.3",
     "jest": "^29.7.0",
